Deleting fix Removed extra modal-component
